+++ lib/l2vpn.h
@@ -119,7 +119,7 @@ extern const struct frr_yang_module_info frr_l2vpn_cli_info;
 extern void l2vpn_cli_init(void);
 extern void l2vpn_init(void);
 extern void l2vpn_init_new(bool in_backend);
-extern const char *l2vpn_pw_error_code(uint8_t status);
+extern const char *l2vpn_pw_error_code(uint32_t status);
 
 struct l2vpn *l2vpn_new(const char *name);
 struct l2vpn *l2vpn_find(struct l2vpn_head *conf, const char *name, int type);
